+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## [Unreleased]
 
+## [0.4.0]
+### atualiz
+
+mudulariza - bulma, agora h-bulma.js
+remove tag personalizadas de hypertag
+
+
 ## [0.3.0]
 ### inclui
 - h.use('bulma') para gerar tag personalizadas